backend

+++ resources/membership/routes.py
@@ -1,4 +1,5 @@
 from flask import request
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from uuid import uuid4
 from flask.views import MethodView
 from flask_smorest import abort
@@ -16,22 +17,23 @@ class Membership(MethodView):
   def get(self, membership_id):
     membership = MembershipModel.query.get(membership_id)
     if membership:
-      print(membership.author)
       return membership 
     abort(400, message='Invalid Membership')
 
+  @jwt_required
   @bp.arguments(MembershipSchema)
   def put(self, membership_data ,membership_id):
     membership = MembershipModel.query.get(membership_id)
-    if membership:
+    if membership and membership.user_id == get_jwt_identity():
       membership.body = membership_data['body']
       membership.commit()
       return {'message': 'membership updated'}, 201
     return {'message': "Invalid Membership Id"}, 400
 
+  @jwt_required
   def delete(self, membership_id):
     membership = MembershipModel.query.get(membership_id)
-    if membership:
+    if membership and membership.user_id == get_jwt_identity():
       membership.delete()
       return {"message": "Membership Deleted"}, 202
     return {'message':"Invalid Membership"}, 400
@@ -43,12 +45,13 @@ class MembershipList(MethodView):
   def get(self):
     return MembershipModel.query.all()
   
+  @jwt_required
   @bp.arguments(MembershipSchema)
   def membership(self, membership_data):
     try:
       membership = MembershipModel()
-      membership.user_id = membership_data['user_id']
-      membership.body = membership_data['membership']
+      membership.user_id = get_jwt_identity()
+      membership.body = membership_data['body']
       membership.commit()
       return { 'message': "Membership Created" }, 201
     except:
